session_helperの実装をcontrollers/concerns/session_util に移行
